refactor(TextChannel): new base class
