+++ src/structures/NewsChannel.js
@@ -1,20 +1,13 @@
 'use strict';
 
-const TextChannel = require('./TextChannel');
+const BaseGuildTextChannel = require('./BaseGuildTextChannel');
 const { Error } = require('../errors');
 
 /**
  * Represents a guild news channel on Discord.
- * @extends {TextChannel}
+ * @extends {BaseGuildTextChannel}
  */
-class NewsChannel extends TextChannel {
-  _patch(data) {
-    super._patch(data);
-
-    // News channels don't have a rate limit per user, remove it
-    this.rateLimitPerUser = undefined;
-  }
-
+class NewsChannel extends BaseGuildTextChannel {
   /**
    * Adds the target to this channel's followers.
    * @param {GuildChannelResolvable} channel The channel where the webhook should be created
